requirements.txt
Django==5.0.1
python-dotenv==1.0.0
Pillow
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-filter==23.5
django-import-export==4.0.5
openpyxl==3.1.2
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
django-cors-headers==4.3.1

django-storages==1.14.2
boto3==1.34.11
